requirements.txt
keras==2.12.0
lifelines==0.27.4
matplotlib==3.5.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
scipy==1.10.1
skimage==0.0
tensorflow==2.12.0
tensorflow_macos==2.12.0
